# docs-site/src/content/docs/api/openapi.json
{
  "components": {
    "schemas": {
      "EmbedRequest": {
        "description": "Request body for embedding generation.",
        "properties": {
          "texts": {
            "description": "One or more texts to generate embeddings for.",
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Texts",
            "type": "array"
          }
        },
        "required": [
          "texts"
        ],
        "title": "EmbedRequest",
        "type": "object"
      },
      "EmbedResponse": {
        "description": "Embedding generation response body.",
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "dimension": {
            "title": "Dimension",
            "type": "integer"
          },
          "elapsed_seconds": {
            "title": "Elapsed Seconds",
            "type": "number"
          },
          "embeddings": {
            "items": {
              "items": {
                "type": "number"
              },
              "type": "array"
            },
            "title": "Embeddings",
            "type": "array"
          },
          "model_name": {
            "title": "Model Name",
            "type": "string"
          }
        },
        "required": [
          "embeddings",
          "model_name",
          "dimension",
          "count",
          "elapsed_seconds"
        ],
        "title": "EmbedResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HealthResponse": {
        "description": "Health check response body.",
        "properties": {
          "model_loaded": {
            "title": "Model Loaded",
            "type": "boolean"
          },
          "model_name": {
            "title": "Model Name",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "version": {
            "default": "0.1.0",
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "status",
          "model_loaded",
          "model_name"
        ],
        "title": "HealthResponse",
        "type": "object"
      },
      "ProcessRequest": {
        "description": "Request body for inline or file pipeline processing.",
        "properties": {
          "generate_embeddings": {
            "default": false,
            "description": "Whether to generate embeddings.",
            "title": "Generate Embeddings",
            "type": "boolean"
          },
          "input": {
            "description": "Input text or file path to process.",
            "minLength": 1,
            "title": "Input",
            "type": "string"
          },
          "source": {
            "default": "legislation",
            "description": "Corpus source type.",
            "pattern": "^(legislation|hansard)$",
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "input"
        ],
        "title": "ProcessRequest",
        "type": "object"
      },
      "ProcessResponse": {
        "description": "Pipeline processing response body.",
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "elapsed_seconds": {
            "title": "Elapsed Seconds",
            "type": "number"
          },
          "output_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Path"
          },
          "records": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Records",
            "type": "array"
          },
          "source": {
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "records",
          "source",
          "count",
          "elapsed_seconds"
        ],
        "title": "ProcessResponse",
        "type": "object"
      },
      "SearchRequest": {
        "description": "Request body for semantic search.",
        "properties": {
          "db_path": {
            "default": "./lancedb_data",
            "description": "Path to the LanceDB database directory.",
            "title": "Db Path",
            "type": "string"
          },
          "query": {
            "description": "Natural-language search query.",
            "minLength": 1,
            "title": "Query",
            "type": "string"
          },
          "top_k": {
            "default": 5,
            "description": "Number of results to return.",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Top K",
            "type": "integer"
          }
        },
        "required": [
          "query"
        ],
        "title": "SearchRequest",
        "type": "object"
      },
      "SearchResponse": {
        "description": "Semantic search response body.",
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "elapsed_seconds": {
            "title": "Elapsed Seconds",
            "type": "number"
          },
          "query": {
            "title": "Query",
            "type": "string"
          },
          "results": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Results",
            "type": "array"
          }
        },
        "required": [
          "results",
          "query",
          "count",
          "elapsed_seconds"
        ],
        "title": "SearchResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "HTTP API for the NLP Policy NZ pipeline. Provides embedding generation, semantic search, and full pipeline processing.",
    "title": "nlp-policy-nz Inference API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/embed": {
      "post": {
        "description": "Generate embeddings for one or more input texts.",
        "operationId": "embed_embed_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmbedRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmbedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate embeddings"
      }
    },
    "/health": {
      "get": {
        "description": "Return API health and lazy model-load state.",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health check"
      }
    },
    "/process": {
      "post": {
        "description": "Run pipeline processing against a path or inline text.",
        "operationId": "process_process_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProcessRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run full pipeline"
      }
    },
    "/search": {
      "post": {
        "description": "Run semantic vector search over the configured index.",
        "operationId": "search_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Semantic search"
      }
    }
  }
}